requirements.txt
flask==3.0.0
flask-session==0.5.0
flask-limiter==3.5.0
requests==2.31.0
python-dotenv==1.0.0
